feat: enable tapping on feed tab to see user details screen
