+++ qaul_ui/lib/screens/home/tabs/tab.dart
@@ -13,13 +13,14 @@ import 'package:qaul_ui/decorators/disabled_state_decorator.dart';
 import 'package:qaul_ui/decorators/loading_decorator.dart';
 import 'package:qaul_ui/screens/chat/chat.dart';
 import 'package:qaul_ui/screens/chat/models/text_message.dart';
-import 'package:qaul_ui/widgets/default_back_button.dart';
 import 'package:qaul_ui/decorators/empty_state_text_decorator.dart';
 import 'package:qaul_ui/widgets/user_avatar.dart';
 import 'package:qaul_rpc/qaul_rpc.dart';
 import 'package:utils/utils.dart';
 import 'package:uuid/uuid.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+
+import '../user_details_screen.dart';
 
 part 'chat_tab.dart';
